requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
WTForms==3.2.1

